Bugfix variable type

+++ vorlesung/projekt/lenaflomarco/src/UserManager/UserDatabase.go
@@ -93,7 +93,7 @@ func ReadUser(email string) (record UserRecord, present bool, err error) {
 	return
 }
 
-func getNextUID() (id int) {
+func getNextUID() (id int16) {
 	id = 0
 	for _, value := range managersUserStorage.UserMap {
 		if(value.UID > id) {
